analytics-service/requirements.txt
Django
djangorestframework
django-cors-headers
pandas
numpy
matplotlib
seaborn
wordcloud
scikit-learn
textstat
requests
python-dotenv
pymongo
celery
redis